+++ packages/desktop/src/renderer/components/LoadingPanel/loadingPanel.test.tsx
@@ -17,7 +17,7 @@ describe('CreateUsernameModal', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1hk9oxe-MuiModal-root"
+          class="Modalroot MuiModal-root css-1voaj9u-MuiModal-root"
           role="presentation"
         >
           <div