+++ packages/desktop/src/shared/static.ts
@@ -3,6 +3,7 @@ import mirrorKey from 'keymirror'
 export const DOMAIN = 'handlers.zbay.app'
 export const LOG_ENDPOINT = 'https://handlers.zbay.app/email'
 export const DEV_DATA_DIR = 'Quietdev'
+export const DATA_DIR = 'Quiet'
 
 export const actionTypes = mirrorKey({
   SET_APP_VERSION: undefined