+++ packages/mobile/android/app/src/main/java/com/quietmobile/Scheme/WebsocketConnectionPayload.kt
@@ -1,0 +1,5 @@
+package com.quietmobile.Scheme
+
+data class WebsocketConnectionPayload (
+    val dataPort: Int
+)